[ReactDnd] change NativeTypes to use string typing (#27)
The NativeTypes exposed by the html5 backend were previously typed as ints through an enum. However these types are meant to be used as drag source types by function such as DragSource() and DropTarget(), which expect string types. This pr changes the NativeTypes declaration to use strings instead of ints.

* updated NativeTypes from int enum to string typing to match usage spec

* whitespace fix

* changed indentation style

* using namespace instead